Adding event handler for special processing of requests.
Certain requests must return a value to the client, but don't necessarily need to
block the server to protect resources. An event handler was was created to process
these request asynchronously while still blocking the client. The results are then
further processed and handled as replies to be sent back to the database server.
Pings and logouts are now as asynchronous as possible.